refactor: DRY the simple success responses.
