+++ src/messages/responses/cache-set-remove-elements.ts
@@ -1,14 +1,1 @@
-import {ResponseBase, ResponseError, ResponseSuccess} from './response-base';
-import {SdkError} from '../../errors/errors';
-
-export abstract class Response extends ResponseBase {}
-
-class _Success extends Response {}
-export class Success extends ResponseSuccess(_Success) {}
-
-class _Error extends Response {
-  constructor(protected _innerException: SdkError) {
-    super();
-  }
-}
-export class Error extends ResponseError(_Error) {}
+export * from './response-simple-success';
