+++ src/messages/responses/create-cache.ts
@@ -1,16 +1,4 @@
-import {SdkError} from '../../errors/errors';
-import {ResponseBase, ResponseError, ResponseSuccess} from './response-base';
-
-export abstract class Response extends ResponseBase {}
-
-class _Success extends Response {}
-export class Success extends ResponseSuccess(_Success) {}
+import {Response} from './response-simple-success';
+export * from './response-simple-success';
 
 export class AlreadyExists extends Response {}
-
-class _Error extends Response {
-  constructor(protected _innerException: SdkError) {
-    super();
-  }
-}
-export class Error extends ResponseError(_Error) {}
