+++ src/messages/responses/response-simple-success.ts
@@ -1,0 +1,14 @@
+import {SdkError} from '../../errors/errors';
+import {ResponseBase, ResponseError, ResponseSuccess} from './response-base';
+
+export abstract class Response extends ResponseBase {}
+
+class _Success extends Response {}
+export class Success extends ResponseSuccess(_Success) {}
+
+class _Error extends Response {
+  constructor(protected _innerException: SdkError) {
+    super();
+  }
+}
+export class Error extends ResponseError(_Error) {}
